Added check for existence.
Former-commit-id: 0abad42fa22002256ed5d6f2967478ebcda2b2d7
Former-commit-id: 710ce2ca92294ec43cbd8c3449e7409c09782b41

+++ test/src/main/java/org/icgc/dcc/mongodb/MongoExporter.java
@@ -59,7 +59,9 @@ public class MongoExporter extends BaseMongoImportExport {
   @SneakyThrows
   private void exportCollection(File collectionFile, MongoCollection collection) {
     Files.createParentDirs(collectionFile);
-    checkState(collectionFile.delete(), "Collection file not deleted: %s", collectionFile);
+    if(collectionFile.exists()) {
+      checkState(collectionFile.delete(), "Collection file not deleted: %s", collectionFile);
+    }
     checkState(collectionFile.createNewFile(), "Collection file not created: %s", collectionFile);
 
     ObjectMapper mapper = new ObjectMapper();
